auth-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
httpx==0.25.2
qdrant-client==1.7.0